add account test bed
